extraction/requirements.txt
playwright
asyncio
google-cloud-aiplatform
google-cloud-storage
